Moved profanity filter check and block list
Runs the profanity check and block list only if a condition that would trigger the bot actually runs. While technically not a "bug" in the traditional sense, this will help prevent unnecessary spamming in the console if the block list or profanity was to be triggered by every message for example.